♻️REFACTOR : slice 분리 및 get selector 호출 방식 변경
