+++ package.json
@@ -41,6 +41,8 @@
     "webpack-dev-server": "^4.8.1"
   },
   "dependencies": {
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
     "@reduxjs/toolkit": "^1.9.5",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
