+++ src/App.jsx
@@ -1,7 +1,6 @@
 import { Switch, Route, Link } from 'react-router-dom';
 
 import { useDispatch } from 'react-redux';
-
 import HomePage from './pages/HomePage';
 import AboutPage from './pages/AboutPage';
 import LoginPage from './pages/LoginPage';
@@ -9,9 +8,10 @@ import RestaurantsPage from './pages/RestaurantsPage';
 import RestaurantPage from './pages/RestaurantPage';
 import NotFoundPage from './pages/NotFoundPage';
 
-import { setAccessToken } from './redux/slice';
+import { setAccessToken } from './redux/slice/loginSlice';
 
 import { loadItem } from './services/storage';
+import styled from '@emotion/styled';
 
 export default function App() {
   const dispatch = useDispatch();
@@ -23,11 +23,11 @@ export default function App() {
 
   return (
     <div>
-      <header>
+      <Header>
         <h1>
           <Link to="/">헤더 영역</Link>
         </h1>
-      </header>
+      </Header>
       <Switch>
         <Route exact path="/" component={HomePage} />
         <Route path="/about" component={AboutPage} />
@@ -39,3 +39,8 @@ export default function App() {
     </div>
   );
 }
+
+const Header = styled.header({
+  width: '90%',
+  margin: '0 auto',
+});
