+++ src/components/LoginFormContainer.jsx
@@ -3,16 +3,24 @@ import { useDispatch, useSelector } from 'react-redux';
 import LoginForm from './LoginForm';
 import LogoutForm from './LogoutForm';
 
-import { changeLoginField, requestLogin, logout } from '../redux/slice';
+import {
+  changeLoginField,
+  requestLogin,
+  logout,
+} from '../redux/slice/loginSlice';
 
 import { get } from '../utils/utils';
 
 export default function LoginFormContainer() {
   const dispatch = useDispatch();
 
-  const loginFields = useSelector(get('loginFields'));
-  const accessToken = useSelector(get('accessToken'));
-
+  const loginFields = useSelector(
+    get({ sliceName: 'loginSlice', key: 'loginFields' })
+  );
+  const accessToken = useSelector(
+    get({ sliceName: 'loginSlice', key: 'accessToken' })
+  );
+  console.log(loginFields);
   const handleChange = ({ name, value }) => {
     dispatch(changeLoginField({ name, value }));
   };
