+++ src/components/RestaurantsContainer.jsx
@@ -3,7 +3,9 @@ import { useSelector } from 'react-redux';
 import { get } from '../utils/utils';
 
 export default function RestaurantsContainer({ onClickRestaurant }) {
-  const restaurants = useSelector(get('restaurants'));
+  const restaurants = useSelector(
+    get({ sliceName: 'restaurantSlice', key: 'restaurants' })
+  );
 
   function handleClick(restaurant) {
     return (event) => {
