+++ src/pages/RestaurantsPage.jsx
@@ -8,7 +8,7 @@ import RegionsContainer from '../components/RegionsContainer';
 import CategoriesContainer from '../components/CategoriesContainer';
 import RestaurantsContainer from '../components/RestaurantsContainer';
 
-import { loadInitialData } from '../redux/slice';
+import { loadInitialData } from '../redux/slice/restaurantSlice';
 
 export default function RestaurantsPage() {
   const history = useHistory();
