+++ src/services/api.js
@@ -13,16 +13,17 @@ export async function fetchCategories() {
 }
 
 export async function fetchRestaurants({ regionName, categoryId }) {
-  const url = 'https://eatgo-customer-api.ahastudio.com/restaurants'
-    + `?region=${regionName}&category=${categoryId}`;
+  const url =
+    'https://eatgo-customer-api.ahastudio.com/restaurants' +
+    `?region=${regionName}&category=${categoryId}`;
   const response = await fetch(url);
   const data = await response.json();
   return data;
 }
 
 export async function fetchRestaurant({ restaurantId }) {
-  const url = 'https://eatgo-customer-api.ahastudio.com'
-    + `/restaurants/${restaurantId}`;
+  const url =
+    'https://eatgo-customer-api.ahastudio.com' + `/restaurants/${restaurantId}`;
   const response = await fetch(url);
   const data = await response.json();
   return data;
@@ -42,10 +43,14 @@ export async function postLogin({ email, password }) {
 }
 
 export async function postReview({
-  accessToken, restaurantId, score, description,
+  accessToken,
+  restaurantId,
+  score,
+  description,
 }) {
-  const url = 'https://eatgo-customer-api.ahastudio.com'
-    + `/restaurants/${restaurantId}/reviews`;
+  const url =
+    'https://eatgo-customer-api.ahastudio.com' +
+    `/restaurants/${restaurantId}/reviews`;
 
   const response = await fetch(url, {
     method: 'POST',
